refactor: parameter validation for utils
