+++ services/roles/account.js
@@ -1,16 +1,15 @@
 const bcrypt = require('bcryptjs')
 const { APIError } = require('../../helpers/api-error')
-const { status, code } = require('../../config/result-status-table').errorTable
+const { code } = require('../../config/result-status-table').errorTable
 const { AuthToolKit } = require('../../utils/auth-tool-kit')
 const { FileUploader } = require('../../middlewares/file-uploader')
-const { ParameterValidationKit } = require('../../utils/parameter-validation-kit')
+const { AccountToolKit } = require('../../utils/account-tool-kit')
 
 const { User } = require('../../db/models')
-
-const { blackListRoleIn } = require('../../config/app').generalConfig
 const {
   DEFAULT_BCRYPT_COMPLEXITY,
-  DEL_OPERATION_CODE
+  DEL_OPERATION_CODE,
+  DEFAULT_AVATAR
 } = require('../../config/app').service.accountService
 
 class AccountService {
@@ -20,23 +19,14 @@ class AccountService {
 
   async login(req, cb) {
     try {
-      const { account, password } = req.body
       const type = this.serviceType
-      const { isFilledField } = ParameterValidationKit
+      const { error, result } = await AccountToolKit.loginFormValidate(req, type)
 
-      if (!isFilledField(account) || !isFilledField(password)) {
-        return cb(new APIError({ code: code.FORBIDDEN, status, message: '未填寫完所有欄位' }))
+      if (error) {
+        return cb(new APIError({ code: result.code, data: result.data, message: result.message }))
       }
-      const user = await User.findOne({ where: { account } })
-
-      if (!user || blackListRoleIn[type].includes(user.role)) {
-        return cb(new APIError({ code: code.NOTFOUND, status, message: '帳號不存在' }))
-      }
-
-      if (!bcrypt.compareSync(password, user.password)) {
-        return cb(new APIError({ code: code.FORBIDDEN, status, message: '帳號或密碼不正確' }))
-      }
-      const resultUser = user.toJSON()
+      console.log('user', result.data)
+      const resultUser = result.data.toJSON()
       const accessToken = AuthToolKit.generateAccessToken(resultUser)
       delete resultUser.password
       return cb(null, { accessToken, ...resultUser }, '登入成功')
@@ -47,7 +37,7 @@ class AccountService {
 
   async register(req, cb) {
     try {
-      const message = await ParameterValidationKit.registerFormValidate(req)
+      const message = await AccountToolKit.registerFormValidate(req)
       if (message.length > 0) {
         return cb(new APIError({ code: code.BADREQUEST, message, data: req.body }))
       }
@@ -71,7 +61,7 @@ class AccountService {
   async putSelf(req, cb) {
     try {
       const user = AuthToolKit.getUser(req)
-      const message = await ParameterValidationKit.updateFormValidate(req)
+      const message = await AccountToolKit.updateFormValidate(req)
 
       if (message.length > 0) {
         return cb(new APIError({ code: code.BADREQUEST, message, data: req.body }))
@@ -83,7 +73,7 @@ class AccountService {
       let uploadAvatar = ''
 
       if (avatar === DEL_OPERATION_CODE) {
-        uploadAvatar = 'https://res.cloudinary.com/dqfxgtyoi/image/upload/v1646039874/twitter/project/defaultAvatar_a0hkxw.png'
+        uploadAvatar = DEFAULT_AVATAR
       } else {
         uploadAvatar = file
           ? await FileUploader.upload(file, destType)
